limit case: the user is still connected since yesterday
In that case, if the computer was not rebooted, parentrol must kill the user in order to restore its parameters and allow them to log in today. It means that at midnight the users will be slain even if their session is not active. If someone is working on the commputer, their session will flicker. Oh, well: only I work on the computer at midnight :)